@@ -1,57 +1,26 @@
 {
-  "name": "react-cohort-graph",
-  "version": "1.1.1",
-  "description": "Cohort Analysis Graph with ReactJS",
-  "main": "lib/reactcohortgraph.js",
-  "scripts": {
-    "build": "webpack --env build",
-    "dev": "webpack --progress --colors --watch --env dev --display-error-details",
-    "test": "mocha --compilers js:babel-core/register --colors ./test/*.spec.js",
-    "test:watch": "mocha --compilers js:babel-core/register --colors -w ./test/*.spec.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/apxor/react-cohort-graph.git"
-  },
-  "keywords": [
-    "cohort",
-    "retention",
-    "graph",
-    "cohort",
-    "analysis",
-    "retention",
-    "graph",
-    "react",
-    "graph"
-  ],
-  "author": "Jyothi Babu Araja <arajajyothibabu@gmail.com> (https://arajajyothibabu.github.io/)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/apxor/react-cohort-graph/issues"
-  },
-  "homepage": "https://github.com/apxor/react-cohort-graph#readme",
-  "peerDependencies": {
-    "react": ">=0.14",
-    "react-dom": ">=0.14"
-  },
-  "devDependencies": {
-    "babel-cli": "6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-add-module-exports": "0.2.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "chai": "3.5.0",
-    "mocha": "3.3.0",
-    "react": "15.5.4",
-    "react-dom": "15.5.4",
-    "webpack": "^2.6.1",
-    "webpack-dev-server": "^2.4.5",
-    "yargs": "7.1.0"
-  },
+  "name": "react-cohort-graph-docs",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "prop-types": "15.5.10"
-  }
+    "react": "^16.8.3",
+    "react-cohort-graph": "file:package",
+    "react-dom": "^16.8.3",
+    "react-scripts": "2.1.5"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
